+++ src/lib/audio/audio-loader.svelte
@@ -1,53 +1,53 @@
 <script lang="ts">
 	import {
-		audio_autoplay,
-		audio_current_time,
-		audio_duration,
-		audio_element,
-		audio_ended,
-		audio_loading,
-		audio_muted,
-		audio_paused,
-		audio_src,
-		audio_start_at,
-	} from '$lib/context/audio-internals';
-	import { user_preferences } from '$lib/context/user-preferences';
+		_autoplay,
+		_current_time,
+		_duration,
+		_element,
+		_ended,
+		_loading,
+		_muted,
+		_paused,
+		_src,
+		_start_at,
+	} from '$lib/audio/_private';
+	import { podcast_preferences } from '$lib/preferences';
 	import type { AudioPlayerElement } from '$lib/types';
 	import { load_podcast_state } from '$lib/utility/use-state';
 	import { onMount } from 'svelte';
 
 	// readonly values
 	let element: AudioPlayerElement;
-	let current_time = $audio_start_at;
+	let current_time = $_start_at;
 	let duration = 0;
 	let ended = true;
 	let paused = false;
 
 	// handler values
-	let volume = $user_preferences.volume;
-	let playbackRate = $user_preferences.playback_rate;
+	let volume = $podcast_preferences.volume;
+	let playbackRate = $podcast_preferences.playback_rate;
 	let muted = false;
 
 	// readonly - stores will update when bindings change
-	$: audio_element.set(element);
-	$: audio_current_time.set(current_time);
-	$: audio_duration.set(duration);
-	$: audio_ended.set(ended);
-	$: audio_paused.set(paused);
+	$: _element.set(element);
+	$: _current_time.set(current_time);
+	$: _duration.set(duration);
+	$: _ended.set(ended);
+	$: _paused.set(paused);
 
 	// handlers - when store value changes, the binding will update the audio element
-	$: volume = $user_preferences.volume;
-	$: playbackRate = $user_preferences.playback_rate;
-	$: muted = $audio_muted;
-	$: current_time = $audio_start_at;
+	$: volume = $podcast_preferences.volume;
+	$: playbackRate = $podcast_preferences.playback_rate;
+	$: muted = $_muted;
+	$: current_time = $_start_at;
 
 	onMount(load_podcast_state);
 </script>
 
 <audio
 	bind:this={element}
-	src={$audio_src}
-	autoplay={$audio_autoplay}
+	src={$_src}
+	autoplay={$_autoplay}
 	bind:currentTime={current_time}
 	bind:muted
 	bind:duration
@@ -71,6 +71,6 @@
 		// TODO: reset player state
 		// console.log('emptied', e);
 	}}
-	on:loadeddata={() => audio_loading.set(false)}
-	on:loadstart={() => audio_loading.set(true)}
+	on:loadeddata={() => _loading.set(false)}
+	on:loadstart={() => _loading.set(true)}
 />